Use a standard python project structure
Instead of having the main file outside of a package, and the main
library package called `lib`, move everything under a `tuxkatana`
package. This is in line with standard python practices. This makes
installing the app as a standalone package much easier.

The params yaml files have also been moved to the `tuxkatana` package,
which means we can now refer to them relative to wherever the app is
installed, removing the need to run the app from the same directory
where the params are.